Fixed error when updating software

+++ app/main.ts
@@ -253,14 +253,17 @@ ipcMain.on('getVersion', (event, data) => {
 
 
 function clearDirectory(directory){
-  fs.readdir(directory, (err, files) => {
-    if (err) throw err;
+  if(fs.existsSync(directory)){
+    fs.readdir(directory, (err, files) => {
+      if (err) throw err;
+    
+      for (const file of files) {
+        fs.unlink(path.join(directory, file), err => {
+          if (err) throw err;
+        });
+      }
+    });
+  }
   
-    for (const file of files) {
-      fs.unlink(path.join(directory, file), err => {
-        if (err) throw err;
-      });
-    }
-  });
 }
